constuctor(gh-repos-repo): set slots NAME, OWNER from arg NEWNAME
Set slot NAME and OWNER from the constructor argument NEWNAME if that
has the from "<OWNER>/<NAME>".  Actually, if the new slot FULL-NAME is
set in the constructor arguments use that instead.

This allows writing less to get an gh-repos-repo object suitable for
any api request.  Instead of

  (gh-repos-repo "repo" :name "<NAME>"
                 :owner (gh-user "owner" :login "<OWNER>"))

which is to long, simply

  (gh-repos-repo "repo" :full-name "<OWNER>/<NAME>")

or even

  (gh-repos-repo "<OWNER>/<NAME>")

can be used now.