gh-sanitize-content: new function
This function removes all non-unicode characters from arbitrary content. This
can be used to ensure that the API call will not fail server-side to parse the
request.

Since this is a lossy operation, warn the user.

Use-case: buffers that we want to gist can contain arbitrary content, but we
still need to encode it.

This closes defunkt/gist.el#86